# change dict hash key func for performance and maintain insertion order