--- a/src/frozenset.js
+++ b/src/frozenset.js
@@ -370,7 +370,6 @@ Sk.builtin.frozenset.prototype["copy"] = new Sk.builtin.func(function (self) {
 
 Sk.exportSymbol("Sk.builtin.frozenset", Sk.builtin.frozenset);
 
-
 Sk.builtin.frozenset.prototype.__contains__ = new Sk.builtin.func(function(self, item) {
     Sk.builtin.pyCheckArgsLen("__contains__", arguments.length, 2, 2);
     return new Sk.builtin.bool(self.sq$contains(item));