this pr aims to improve dictionary lookups. 

It changes the hash function to do the following:
```javascript
function kf(key) {
    let key_hash = key.$savedKeyHash_;
    if (key_hash !== undefined) {
        return key_hash;
    } else if (key.ob$type === Sk.builtin.str) {
        key_hash = key.$jsstr().replace(/^[0-9!#_]/, "!$&"); // avoid numbers and clashes 
        key.$savedKeyHash_ = key_hash;
        return key_hash;
    } else if (typeof key === "string") {
        // temporary while sysModules allows javascript strings as keys to python dicts
        return key.replace(/^[0-9!#_]/, "!$&");
    }
    key_hash = Sk.builtin.hash(key).v; // builtin.hash returns an int
    key.$savedKeyHash_ = key_hash;
    return key_hash;
};
```


---

# How it works
```python
d = {1:0, 1.1: 1, 1.2: 2, 'a' : 3, 'b': 4, 2: 5}
# the hash function in skulpt will see collisions for the first three keys
```

Is mapped to... 
```javascript
this: {//dict
  buckets:  // keep track of collisions 
    1: [{lhs: int_(1), rhs: int_(0)}, {lhs: float_(1.1), rhs: int_(1)}, {lhs: float_(1.2), rhs: int_(2)}] //keep track of collisions
    2: [{lhs: int_(2), rhs: int_(5)}] //keep track of collision
  entries:  // preserves insertion order and provide a fast path for str entries...
    #1_0: {lhs: int_(1),     rhs: int_(0)} // non string (int_(1)
    #1_1: {lhs: float_(1.1), rhs: int_(1)} // non string
    #1_2: {lhs: float_(1.2), rhs: int_(2)} // non string
     a:   {lhs: str('a'),    rhs: int_(3)} //string 
     b:   {lhs: str('b'),    rhs: int_(4)} // string 
    #2_0: {lhs: int_(2),     rhs: int_(5)} 
  size: 6
}
```

---

the rationale is two-fold:
- insertion order
  - python 3.6 preserves insertion order
  - javascript Objects preserve insertion order for strings
  - so using strings as keys in will also preserve insertion order
- iteration
  - over a dictionary is now as simple as

```javascript
for (let key_hash in this.entries) {
  item = this.entries[key_hash];
  key = item.lhs;
  val = item.rhs;
}
```

---
### tests

I edited print tests where order changed as the result of the new algorithm


---

### A note on sets
sets will now likely preserve insertion order because they use dict.js under the hood, 
I don't see this as a problem since the user should not expect sets to be ordered... 

```python
set([5,4,3,2,1,6,7])
# before {1, 2, 3, 4, 5, 6, 7}
# now {5, 4, 3, 2, 1, 6, 7}

```

---

### dealing with deleted items

replace the `item` in `bucket` with `undefined`

e.g. 
```python
d = {1:0, 1.1: 1, 1.2: 2}
del d[1]
```
the dict object will look like

```javascript
// deleted the zeroth item with hash value = 1
this: {//dict
  buckets:  // undefined signals a free slot
    1: [**undefined**, {lhs: float_(1.1), rhs: int_(1)}, {lhs: float_(1.2), rhs: int_(2)}] //keep track of collisions
  entries: 
    #1_1: {lhs: float_(1.1), rhs: int_(1)} // non string
    #1_2: {lhs: float_(1.2), rhs: int_(2)} // non string
  size: 2
```
The `undefined` value signifies there is a `free slot` and this will get filled upon the next collision with the hash = `1`.

e.g.
```python
d[1.3] = 3
# skulpt hash function maps 1.3 to the hash 1 i.e a collision
```

```javascript
this: {//dict
  buckets:  // inserted into free slot
    1: [{lhs: float_(1.3), rhs: int_(3)}, {lhs: float_(1.1), rhs: int_(1)}, {lhs: float_(1.2), rhs: int_(2)}] 
  entries: // preserves insertion order 
    #1_1: {lhs: float_(1.1), rhs: int_(1)} // non string
    #1_2: {lhs: float_(1.2), rhs: int_(2)} // non string
    #1_0: {lhs: float_(1.3), rhs: int_(3)} 
  size: 3
```

---
### does it improve speed?

Yes. 

```
fields.py 100 Ticks 10 loops in npm repl
this branch avg: 2.722899961471557
Master avg: 3.401200008392334
20% improvement
```

update (latest version of master approx 6-10%)