@rixner 2020-05-04
This is a reasonably invasive change for no performance improvement.  So, it appears that it comes down to whether we want to maintain insertion order?  Is this the "best" way to do that?