@albertjan 2020-05-24
I ran this through field.py, and I saw a 3% perf improvement see numbers below: 

master:
```
external.        - field.py
Run time: 1.235s - time elapsed: 0.9880001544952393
Run time: 1.223s - time elapsed: 0.9809999465942383
Run time: 1.226s - time elapsed: 0.9739999771118164
Run time: 1.241s - time elapsed: 0.9850001335144043
Run time: 1.257s - time elapsed: 1.000999927520752
Run time: 1.249s - time elapsed: 1.000999927520752
Run time: 1.202s - time elapsed: 0.9590001106262207
Run time: 1.212s - time elapsed: 0.9590001106262207
Run time: 1.208s - time elapsed: 0.9620001316070557
Run time: 1.206s - time elapsed: 0.9560000896453857
avg.    : 1.2259 - avg         : 0.9766000509 
```

this branch
```
external.        - field.py
Run time: 1.18s  - time elapsed: 0.9430000782012939
Run time: 1.252s - time elapsed: 1.013999938964844
Run time: 1.186s - time elapsed: 0.9469997882843018
Run time: 1.18s  - time elapsed: 0.937000036239624
Run time: 1.181s - time elapsed: 0.9409999847412109
Run time: 1.193s - time elapsed: 0.9340000152587891
Run time: 1.182s - time elapsed: 0.9389998912811279
Run time: 1.192s - time elapsed: 0.9500000476837158
Run time: 1.175s - time elapsed: 0.937000036239624
Run time: 1.179s - time elapsed: 0.940000057220459
avg.    : 1.19   - avg         : 0.9481999874
```