@s-cork 2020-05-26
@rixner  worth noting that my `$all_keys` function is really just the dict version of `sk$asarray` (#1054) which I extracted from the logic in creating the `dict_key_iterator`.  

---
side note and possible to add to this pr or to a separate one:

I think that it is worth considering preventing JS objects from being keys/values of a Py dict. 
which is that weird hack I have for the `hashkey function` :
```javascript
// because we sometimes let js strings be keys
else if (typeof key === "string") {
        return "_" + key;
    }
```

We allow this in `Sk.sysmodules = new Sk.builtin.dict;` where we pass in JS strings instead of Py strings. 

Which then causes issues since the dict module uses `Sk.misceval.richCompareBool` to compare keys to the bucket LHS items...  and the documentation for `Sk.misceval.richCompareBool` states that it should only be used with python objects. 😞 

---

is 3% a good (enough) speed improvement?

---

is it worth having a set of generic speed test programs that can be run? So that we can easily compare speeds using the same programs and the differences would be branches and systems.