@albertjan 2020-05-26
> We allow this in Sk.sysmodules = new Sk.builtin.dict; where we pass in JS strings instead of Py strings.

that's weird and I agree we shouldn't allow js objects to live in our dict.

> is 3% a good (enough) speed improvement?

I think that's pretty good @meredydd will have a better view on this, but I think the really big savings have all been done so what remains are these kinds of improvements.

> is it worth having a set of generic speed test programs that can be run? So that we can easily compare speeds using the same programs and the differences would be branches and systems.

Yeah it would be good to formalise these, we've all been running @rixner's field.py script and I've also been running a slightly modified version of pystone.py. We used to have something like that, it would run a script 10 times and collect run time data, @mariechatfield made that. But that disappeared when we moved away from `skulpt.py` to `npm run *`.

[pystone-field.zip](https://github.com/skulpt/skulpt/files/4680807/pystone-field.zip)