@s-cork 2020-05-26
i've added the `Sk.sysModules` change here and as it's own separate pr if this doesn't get pulled. #1098