@s-cork 2020-06-04
Update: I came up with a significantly better hash function than the initial proposal in this pr. 

I've amended accordingly and see `20%` speed improvements on `fields.py` (figures in the pr description)

This new proposal also simplifies dictionary iteration significantly from a JS perspective.