@albertjan 2020-06-04
This is dependent on the the sys modules PR isn't it? Did you push this new hash function? 😄