@s-cork 2020-06-04
This PR includes the `Sys Modules` change - so yes totally dependent on the approval of that pr. 

re hash function - the `Sk.builtin.hash` remains the same. it's the `dict.js` `hash_function (var kf =)` which is different.  It's on line 75 in `dict.js`