@s-cork 2020-08-05
I've reverted the `sysModules` change in this pr so that it is no longer a breaking change.