@s-cork 2020-08-05
I've also added `sk$asarray` to `Sk.builtin.dict.prototype`