@s-cork 2020-08-13
@meredydd - I did another speed test with `fields.py` - I think the sensitivity of the `Sk` object pre #1143 impacted the figures in the description. (Which were valid at the point i made them)
I'm now getting around 10% `2.79 seconds (master) vs 2.49 seconds (this branch)` (200 TICKS)